fix: describe interceptor providers in the CoreModule
